pyproject.toml
[tool.poetry]
name = "ai-coversation"
version = "0.1.0"
description = ""
authors = ["Ryan Lee <lee98064@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.75.0"
flask = "^3.1.0"
flask-socketio = "^5.5.1"
flask-cors = "^5.0.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
